day-91-2026Apr6-agentic-ai/phidata-langchain/requirements.txt
streamlit
phidata
langchain
langchain-community
langchain-core
langchain-openai
langgraph
openai
yfinance
